new focus on tmux managment ready; kept open current path as option

+++ workflow/kill-session.rb
@@ -1,10 +1,14 @@
 #!/usr/bin/env ruby
 # encoding: utf-8
 
-name = ARGV[0].split('|')[1]
+name = ARGV[0].split('|')[2]
 
 sessions = `/usr/local/bin/tmux list-sessions -F "#S"`.split "\n"
 
+unless sessions.include? name
+  `/usr/bin/afplay /System/Library/Sounds/Funk.aiff`
+  exit 1
+end
 exit 1 unless sessions.include? name
 
 `/usr/local/bin/tmux kill-session -t "#{name}"`
